+++ api/routes/tracks.ts
@@ -1,24 +1,23 @@
 import express from 'express';
-import mongoose from 'mongoose';
+import mongoose, { Types } from 'mongoose';
 
 import auth, { RequestWithUser } from '../middleware/auth';
 import Track from '../models/Track';
 import Album from '../models/Album';
-import { AlbumI, TrackI, TrackMutation } from '../types';
 import permit from '../middleware/permit';
+import { AlbumI, TrackI, UserCheck } from '../types';
 
 const tracksRouter = express.Router();
 
-tracksRouter.post('/', auth, async (req, res, next) => {
+tracksRouter.post('/', auth, async (req: RequestWithUser, res, next) => {
   try {
-    const trackData: TrackMutation = {
+    const track = new Track({
       name: req.body.name,
       album: req.body.album,
       duration: req.body.duration,
       albumTrackNumber: req.body.albumTrackNumber,
-    };
-
-    const track = new Track(trackData);
+      user: req.user?._id,
+    });
     await track.save();
 
     res.send(track);
@@ -74,22 +73,42 @@ tracksRouter.get('/', async (req, res, next) => {
   }
 });
 
+const deleteCheck = (user: UserCheck, track: TrackI): boolean => {
+  let _id: Types.ObjectId;
+  try {
+    _id = new Types.ObjectId(user._id);
+  } catch {
+    return false;
+  }
+
+  if (user.role === 'admin') {
+    return true;
+  }
+
+  return user.role === 'user' && !track.isPublished && _id.equals(track.user);
+};
+
 tracksRouter.delete(
   '/:id',
   auth,
-  permit('admin'),
+  permit('admin', 'user'),
   async (req: RequestWithUser, res, next) => {
     try {
       const id = req.params.id;
+      const user = req.user;
 
-      const track = await Track.findById(id);
+      const track: TrackI | null = await Track.findById(id);
 
       if (!track) {
         return res.status(404).send({ error: 'Not found!' });
       }
 
-      await Track.deleteOne({ _id: id });
-      return res.send({ message: 'Track deleted!' });
+      if (user && deleteCheck({ _id: user.id, role: user.role }, track)) {
+        await Track.deleteOne({ _id: id });
+        return res.send({ message: 'Track deleted!' });
+      }
+
+      return res.status(403).send({ error: 'Forbidden' });
     } catch (e) {
       next(e);
     }
